Extract input fetching logic to its own class.
