+++ support/advent_day.rb
@@ -1,8 +1,8 @@
 require 'benchmark'
-require 'net/http'
+
+require_relative 'input_fetcher'
 
 class AdventDay
-  SESSION = ENV['SESSION']
   YEAR = ENV['YEAR']
 
   def self.solve
@@ -18,28 +18,9 @@ class AdventDay
   end
 
   def input
-    input_path = Pathname.new('inputs/'+day_number)
-    input_data = if input_path.exist?
-      File.read(input_path)
-    else
-      download_input
-    end
-    convert_data(input_data)
-  end
-
-  INPUT_BASE_URL = 'https://adventofcode.com'.freeze
-  INPUT_PATH_SCHEME = '/%{year}/day/%{number}/input'.freeze
-
-  def download_input
-    raise "Cannot download input without a session cookie" unless SESSION
-    res = Faraday.get(
-      INPUT_BASE_URL + INPUT_PATH_SCHEME % { year: YEAR, number: day_number },
-      nil,
-      { 'Cookie' => "session=#{SESSION}" },
-    )
-    raise "Input doesn't appear to be accessible (yet?)" if res.status == 404
-    File.write('inputs/'+day_number, res.body)
-    res.body
+    return @input if defined?(@input)
+    input_data = InputFetcher.new(day_number, YEAR).get
+    @input ||= convert_data(input_data)
   end
 
   def day_number
